Implement Doctor controller
Enable CrossOrigin for controllets

+++ back-end/feelin/src/main/java/com/feelin/feelin/repo/DoctorRepo.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.ArrayList;
 
-public interface DoctorRepo extends JpaRepository<Doctor, Integer> {
+public interface DoctorRepo extends JpaRepository<Doctor, Long> {
 
 }
